+++ src/NeatMapper/Exceptions/MappingException.cs
@@ -6,8 +6,8 @@ namespace NeatMapper {
 	/// </summary>
 	public sealed class MappingException : TypesException {
 		public MappingException(Exception? exception, (Type From, Type To) types) :
-			base($"An exception was thrown while mapping the types: {types.From.Name} -> {types.To.Name}\n" +
-			$"{types.From.FullName} -> {types.To.FullName}\n" +
+			base($"An exception was thrown while mapping the types: {types.From.Name} -> {types.To.Name}{Environment.NewLine}" +
+			$"{types.From.FullName} -> {types.To.FullName}{Environment.NewLine}" +
 			$"Check the inner exception for details.", exception) { }
 	}
 }