+++ src/NeatMapper/Exceptions/ProjectionException.cs
@@ -6,8 +6,8 @@ namespace NeatMapper {
 	/// </summary>
 	public sealed class ProjectionException : TypesException {
 		public ProjectionException(Exception exception, (Type From, Type To) types) :
-			base($"An exception was thrown while projecting the types: {types.From.Name} -> {types.To.Name}\n" +
-			$"{types.From.FullName} -> {types.To.FullName}\n" +
+			base($"An exception was thrown while projecting the types: {types.From.Name} -> {types.To.Name}{Environment.NewLine}" +
+			$"{types.From.FullName} -> {types.To.FullName}{Environment.NewLine}" +
 			$"Check the inner exception for details", exception) { }
 	}
 }